requirement.txt
absl-py==2.0.0
asgiref==3.7.2
astor==0.8.1
backports.zoneinfo==0.2.1
colorama==0.4.6
cycler==0.11.0
Django==3.2.25
djangorestframework==3.15.1
fonttools==4.38.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.60.0
h5py==2.10.0
importlib-metadata==6.7.0
imutils==0.5.4
joblib==1.3.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
mysqlclient==2.1.0
numpy==1.21.5
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
protobuf==3.19.3
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sqlparse==0.4.4
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-cpu-estimator==1.15.1
termcolor==2.3.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.7.1
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
